+++ efa/efaday.py
@@ -1,6 +1,7 @@
 import datetime as dt
 from typing import Union
 
+import deprecated
 import pandas as pd
 
 from efa import helpers
@@ -95,7 +96,16 @@ class EFADay:
         """Returns the gas day bounday of the EFA day."""
         return self.end_time - dt.timedelta(hours=17)
 
+    @deprecated.deprecated("Use EFADay.from_start_time instead")
     def from_period_start_time(start_time: dt.datetime):
+        """Returns an EFA Day corresponding to a given utc start time."""
+        settlement_date, sp = helpers.sp_from_timestamp(start_time)
+        if sp <= 46:
+            return EFADay(settlement_date)
+        else:
+            return EFADay(settlement_date) + 1
+
+    def from_start_time(start_time: dt.datetime):
         """Returns an EFA Day corresponding to a given utc start time."""
         settlement_date, sp = helpers.sp_from_timestamp(start_time)
         if sp <= 46: